pyproject.toml
[tool.poetry]
name = "taskmanagerpython"
version = "0.1.0"
description = ""
authors = ["Sergey Nesterov <sergioreklamplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
